Add new configuration and start synchronization just once

+++ src/model.cpp
@@ -106,18 +106,17 @@ ErrorCode Model::removeFile()
 
 void Model::startSync()
 {
-    m_syncTimer->start(m_interval, [this]() {
-        m_scanner->scan(m_mainDirectoryPath);
-        auto outputComparing = m_scanner->comparePreviousAndRecentScanning();
-        m_fileSynchronizer->synchronizeAdded(outputComparing.first);
-        m_scanner->scanForChangedDirs(m_mainDirectoryPath);
-        m_fileSynchronizer->synchronizeRemoved(outputComparing.second);
-        m_scanner->scanForChangedDirs(m_mainDirectoryPath);
-    });
+    if(!m_syncStarted){
+        m_syncStarted = true;
+        m_syncTimer->start(m_interval, [this]() {
+            forceSync();
+        });
+    }
 }
 
 void Model::stopSync()
 {
+    m_syncStarted = false;
     m_syncTimer->stop();
 }
 
@@ -138,12 +137,29 @@ fs::path Model::getMainDirectoryPath()
 }
 
 
-void Model::readConfig(){
-    m_serializer = std::make_unique<SerializerToJSON>(); 
-    fs::remove_all(m_mainDirectoryPath);
-    fs::create_directory(m_mainDirectoryPath);
-    auto [dirs, files] = m_serializer->deserialize();
-    std::filesystem::copy(m_mainDirectoryPath / "../configDirectory", m_mainDirectoryPath, std::filesystem::copy_options::recursive);
+void Model::readConfig(std::istream &std_input){
+    std::cout << "Read back-up configuration or add new configuration - Select 1 or 2: \n"
+              << "1. Add new config \n"
+              << "2. Read back-up \n";
+    int selected;
+    std::cin.clear();
+    std_input >> selected;
+    std::cout << "Provided: " << selected << "\n";
+
+    if(selected == 1){
+        addConfig(std::cin);
+    }
+    else if(selected == 2){
+        m_serializer = std::make_unique<SerializerToJSON>(); 
+        fs::remove_all(m_mainDirectoryPath);
+        fs::create_directory(m_mainDirectoryPath);
+        auto [dirs, files] = m_serializer->deserialize();
+        std::filesystem::copy(m_mainDirectoryPath / "../configDirectory", m_mainDirectoryPath, std::filesystem::copy_options::recursive);
+    }
+    else{
+        std::cout << "Provided incorrect value!\n";
+        ViewConsoleUserInterface::waitForButton();
+    }
 }
 
 void Model::saveConfig(){
@@ -156,4 +172,80 @@ void Model::saveConfig(){
     for(auto const& config : configurations){
         config->serialize();
     }    
+}
+
+void Model::addConfig(std::istream &std_input){
+
+    std::cout << "Provide path with configuration: \n";
+    fs::path dirPath;
+    std::cin.clear();
+    std_input >> dirPath;
+    std::cout << "Provided path: " << dirPath << "\nFiles in directory: \n";
+
+    if (!std::filesystem::exists(dirPath)){
+        std::cout << "Incorrect path! Directory not exist! \n";
+        ViewConsoleUserInterface::waitForButton();
+    } 
+    else{
+        for (auto const &fileEntry : fs::recursive_directory_iterator(dirPath))
+        {
+            if (fileEntry.is_regular_file()){
+                std::cout << fileEntry.path().filename() << "\n";
+            }
+        }
+    
+        std::cout << "Provide in JSON format (example: {\"files\": [\"file11.txt\", \"file12.txt\"]}) which files you want to transfer: \n";
+        
+        json fileJSON;
+        std::vector<std::string> Files;
+
+        try{
+            std::cin.clear();
+            std_input >> fileJSON;
+        
+            for (json::iterator it = fileJSON.begin(); it != fileJSON.end(); ++it) {
+                for(auto file : it.value()){
+                    Files.push_back(file);
+                }
+            }
+
+            std::cout << "Provide copy direction from available dirs: \n";
+            for (auto const &dirEntry : fs::directory_iterator(m_mainDirectoryPath))
+            {
+                std::cout << dirEntry.path().filename() << "\n";
+            }
+
+            std::string dir;
+            std::cin.clear();
+            std_input >> dir;
+           
+            if (!std::filesystem::exists(m_mainDirectoryPath / dir)){
+                std::cout << "Incorrect path! Directory not exist! \n";
+                ViewConsoleUserInterface::waitForButton();
+            }
+            else{
+                for(auto file : Files)
+                {
+                    std::cout << file << "\n";
+                    try{
+                        std::filesystem::copy(dirPath / file, m_mainDirectoryPath / dir / file, std::filesystem::copy_options::recursive);
+                    }
+                    catch (std::exception &e){
+                        std::cout << e.what();
+                    }
+                }
+            }
+        }
+        catch (json::parse_error& e)
+        {
+            std::cerr << e.what() << std::endl;
+            ViewConsoleUserInterface::waitForButton();
+
+        }
+    }
+/*
+/Users/pawellos/projects/file_synchronizer/tmpConfig
+{"files": ["file11.txt", "file14.txt"]}
+
+*/        
 }
